Working Front End
Added the functionality to disable the archives on the front end (see
inc/front.php).  Also added a "postype_" and "taxonomy_" prefix on the arrays
of post types and taxonomies respectively to prevent name collissions.

Finally, added an option to allow users to specify what happens when a disabled
archive is encountered.